refactor(admin): one section per page, instead of one scrolling blob
/admin was a single 1,348-line LiveView stacking six sections — funnel,
sandbox spend, billing, users, active sandboxes, admin actions — and its
mount plus a ten-second refresh re-ran every query behind all six no
matter which one was on screen. Reading the user table paid for the
provider-spend attribution and the Stripe overview, forever, every ten
seconds.

The sections are now pages:

  /admin            funnel + one tile per thing that could be wrong
  /admin/users      the table and every per-account lever
  /admin/sandboxes  the live list + spend by provider
  /admin/billing    MRR, status counts, webhook health
  /admin/activity   the privilege trail, paginated rather than the last 25

A shared AdminLive.Shell renders the tab bar; the detail pages under
/admin/users/:id and /admin/conversations/:id render it too with their
section marked, replacing their "← Admin" back link. /admin/finance is
unchanged apart from picking up the same bar.

Two things fall out of the split:

  * Conversations._unsafe_count_sandboxes_admin/0 is built on the same
    query as the listing, so the overview tile and the page it links to
    cannot disagree about what "active" means. Quotas has its own,
    narrower definition and keeps it.
  * Audit._unsafe_page_admin_events/1 pages the whole trail. The overview
    showed the newest 25 with no way to reach the 26th.

format_hours/1 and idle_share/1 move into AdminLive.Helpers; they were
already copied between the index and the finance page.
