+++ apps/fountain/lib/fountain_web/router.ex
@@ -621,7 +621,14 @@ defmodule FountainWeb.Router do
         {FountainWeb.Live.Hooks, :require_authenticated_user},
         {FountainWeb.Live.Hooks, :require_admin}
       ] do
+      # One section per page. `/admin` was a single LiveView stacking all of
+      # these, whose mount and 10s refresh ran every query behind every
+      # section no matter which one was on screen.
       live "/admin", AdminLive.Index, :index
+      live "/admin/users", AdminLive.Users, :index
+      live "/admin/sandboxes", AdminLive.Sandboxes, :index
+      live "/admin/billing", AdminLive.Billing, :index
+      live "/admin/activity", AdminLive.Activity, :index
       # Lives in ee/ with the rest of billing (#472): it is a revenue page.
       live "/admin/finance", Live.AdminFinanceLive, :index
       live "/admin/users/:id", AdminLive.UserDetail, :show
