+++ apps/fountain/test/fountain_web/live/admin_users_live_test.exs
@@ -1,0 +1,647 @@
+defmodule FountainWeb.AdminUsersLiveTest do
+  use FountainWeb.ConnCase, async: true
+
+  import Phoenix.LiveViewTest
+
+  alias Fountain.Accounts
+
+  defp insert_admin(overrides \\ %{}) do
+    user = insert_active_user(overrides)
+    {:ok, admin} = Accounts.update_user_role(user, "admin")
+    admin
+  end
+
+  describe "AdminLive.Users — user list" do
+    test "displays all users", %{conn: conn} do
+      admin = insert_admin()
+      other = insert_active_user()
+      conn = login_user(conn, admin)
+      {:ok, _lv, html} = live(conn, ~p"/admin/users")
+
+      assert html =~ admin.email
+      assert html =~ other.email
+    end
+
+    test "shows onboarding state for each user", %{conn: conn} do
+      admin = insert_admin()
+      conn = login_user(conn, admin)
+      {:ok, _lv, html} = live(conn, ~p"/admin/users")
+
+      assert html =~ "step_1"
+    end
+  end
+
+  describe "AdminLive.Users — toggle_admin" do
+    test "promotes a regular user to admin", %{conn: conn} do
+      admin = insert_admin()
+      target = insert_active_user()
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv |> element("button[phx-value-id='#{target.id}']", "Make admin") |> render_click()
+
+      updated = Accounts.get_user!(target.id)
+      assert updated.role == "admin"
+    end
+
+    test "demotes an admin to regular user", %{conn: conn} do
+      admin = insert_admin()
+      target = insert_admin()
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv |> element("button[phx-value-id='#{target.id}']", "Remove admin") |> render_click()
+
+      updated = Accounts.get_user!(target.id)
+      assert updated.role == "user"
+    end
+  end
+
+  describe "AdminLive.Users — :refresh handle_info" do
+    test "page re-renders on :refresh message without error", %{conn: conn} do
+      admin = insert_admin()
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      # Send the refresh message directly; the page should still render
+      send(lv.pid, :refresh)
+      html = render(lv)
+      assert html =~ "Admin"
+      assert html =~ "Users"
+    end
+
+    test ":refresh picks up newly inserted users", %{conn: conn} do
+      admin = insert_admin()
+      conn = login_user(conn, admin)
+      {:ok, lv, html_before} = live(conn, ~p"/admin/users")
+
+      new_user = insert_active_user()
+
+      send(lv.pid, :refresh)
+      html_after = render(lv)
+
+      refute html_before =~ new_user.email
+      assert html_after =~ new_user.email
+    end
+  end
+
+  describe "AdminLive.Users — user list formatting" do
+    test "shows joined date for each user", %{conn: conn} do
+      admin = insert_admin()
+      conn = login_user(conn, admin)
+      {:ok, _lv, html} = live(conn, ~p"/admin/users")
+
+      # inserted_at is always set so date column should have a year
+      assert html =~ to_string(Date.utc_today().year)
+    end
+
+    test "shows onboarding_completed_at date when set", %{conn: conn} do
+      admin = insert_admin()
+      # Mark onboarding complete via direct Repo update
+      now = DateTime.utc_now() |> DateTime.truncate(:second)
+
+      {:ok, admin_with_date} =
+        Fountain.Repo.update(Ecto.Changeset.change(admin, onboarding_completed_at: now))
+
+      conn = login_user(conn, admin_with_date)
+      {:ok, _lv, html} = live(conn, ~p"/admin/users")
+
+      assert html =~ to_string(now.year)
+    end
+  end
+
+  describe "AdminLive.Users — sandbox concurrency cap" do
+    test "shows each user's active sandbox count against their cap", %{conn: conn} do
+      admin = insert_admin()
+      insert_sandbox(user_id: admin.id, status: "ready")
+      insert_sandbox(user_id: admin.id, status: "pending")
+
+      conn = login_user(conn, admin)
+      {:ok, _lv, html} = live(conn, ~p"/admin/users")
+
+      assert html =~ "Sandboxes"
+      assert html =~ "2 /"
+    end
+
+    test "admin can raise a user's cap", %{conn: conn} do
+      admin = insert_admin()
+      target = insert_active_user()
+
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv
+      |> element("#sandbox-limit-#{target.id}")
+      |> render_submit(%{"user_id" => target.id, "limit" => "25"})
+
+      assert Fountain.Quotas.sandbox_limit(target.id) == 25
+      assert render(lv) =~ "Sandbox limit override set"
+    end
+
+    # A comped account cannot change its own plan — Billing.change_plan/3
+    # refuses — so without an admin door there is no way at all onto the
+    # entitlements of exactly the accounts an operator hand-manages.
+    test "admin can set a user's plan", %{conn: conn} do
+      admin = insert_admin()
+      target = insert_active_user(plan: "solo")
+
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv
+      |> element("#plan-#{target.id}")
+      |> render_change(%{"user_id" => target.id, "plan" => "scale"})
+
+      assert Fountain.Accounts.get_user(target.id).plan == "scale"
+
+      assert Fountain.Quotas.sandbox_limit(target.id) ==
+               Fountain.Plans.concurrent_sandboxes("scale")
+
+      # The privilege trail, not just the effect: record_admin/1 is
+      # best-effort and silently drops an event type missing from its closed
+      # allowlist, which is how admin actions have shipped unaudited before.
+      assert_admin_event(target.id, "admin.plan.changed")
+    end
+
+    test "admin can comp a user's teammate contacts", %{conn: conn} do
+      admin = insert_admin()
+      target = insert_active_user(plan: "team")
+
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv
+      |> element("#comped-contacts-#{target.id}")
+      |> render_submit(%{"user_id" => target.id, "count" => "2"})
+
+      assert Fountain.Accounts.get_user(target.id).comped_contacts == 2
+      assert_admin_event(target.id, "admin.comped_contacts.changed")
+    end
+
+    defp assert_admin_event(target_id, event_type) do
+      types =
+        target_id
+        |> Fountain.Audit._unsafe_list_admin_events_for_target(50)
+        |> Enum.map(& &1.event_type)
+
+      assert event_type in types,
+             "no #{event_type} admin event recorded (got #{inspect(types)}) — " <>
+               "the type is probably missing from AdminEvent's allowlist"
+    end
+
+    test "an empty field clears the override and the cap follows the plan", %{conn: conn} do
+      admin = insert_admin()
+      target = insert_active_user(plan: "team")
+      {:ok, _} = Fountain.Accounts.update_sandbox_limit(target, 25, actor: "admin")
+
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv
+      |> element("#sandbox-limit-#{target.id}")
+      |> render_submit(%{"user_id" => target.id, "limit" => ""})
+
+      assert render(lv) =~ "Override cleared"
+
+      assert Fountain.Quotas.sandbox_limit(target.id) ==
+               Fountain.Plans.fetch!("team").concurrent_sandboxes
+
+      assert Fountain.Accounts.get_user(target.id).sandbox_limit_override == nil
+    end
+
+    test "admin can drop a cap to zero to cut off an abusive tenant", %{conn: conn} do
+      admin = insert_admin()
+      target = insert_active_user()
+
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv
+      |> element("#sandbox-limit-#{target.id}")
+      |> render_submit(%{"user_id" => target.id, "limit" => "0"})
+
+      assert Fountain.Quotas.sandbox_limit(target.id) == 0
+
+      assert {:error, {:sandbox_quota_exceeded, _}} =
+               Fountain.Quotas.check_sandbox_quota(target.id)
+    end
+
+    test "a negative cap is rejected", %{conn: conn} do
+      admin = insert_admin()
+      target = insert_active_user()
+
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv
+      |> element("#sandbox-limit-#{target.id}")
+      |> render_submit(%{"user_id" => target.id, "limit" => "-1"})
+
+      assert Fountain.Quotas.sandbox_limit(target.id) == Fountain.Quotas.default_limit()
+      assert render(lv) =~ "whole number"
+    end
+
+    test "a non-numeric cap is rejected", %{conn: conn} do
+      admin = insert_admin()
+      target = insert_active_user()
+
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv
+      |> element("#sandbox-limit-#{target.id}")
+      |> render_submit(%{"user_id" => target.id, "limit" => "lots"})
+
+      assert Fountain.Quotas.sandbox_limit(target.id) == Fountain.Quotas.default_limit()
+      assert render(lv) =~ "whole number"
+    end
+  end
+
+  describe "AdminLive.Users — billing column" do
+    test "shows subscription status, trial end and a Stripe link", %{conn: conn} do
+      admin = insert_admin()
+
+      user = insert_active_user()
+
+      user =
+        Fountain.Repo.update!(
+          Ecto.Changeset.change(user,
+            subscription_status: "trialing",
+            trial_ends_at: ~U[2027-03-01 00:00:00Z],
+            stripe_customer_id: "cus_admin_test"
+          )
+        )
+
+      conn = login_user(conn, admin)
+      {:ok, _lv, html} = live(conn, ~p"/admin/users")
+
+      assert html =~ "trialing"
+      assert html =~ "ends 2027-03-01"
+      assert html =~ "https://dashboard.stripe.com/customers/#{user.stripe_customer_id}"
+    end
+  end
+
+  describe "AdminLive.Users — extend_trial" do
+    test "extends the trial and records an admin audit event", %{conn: conn} do
+      admin = insert_admin()
+
+      user =
+        Fountain.Repo.update!(
+          Ecto.Changeset.change(insert_active_user(),
+            subscription_status: "canceled",
+            trial_ends_at: nil
+          )
+        )
+
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv
+      |> form("#extend-trial-#{user.id}", %{"days" => "7"})
+      |> render_submit()
+
+      updated = Fountain.Repo.reload!(user)
+      assert updated.subscription_status == "trialing"
+      expected = DateTime.add(DateTime.utc_now(), 7 * 24 * 60 * 60, :second)
+      assert abs(DateTime.diff(updated.trial_ends_at, expected, :second)) < 60
+
+      assert Enum.any?(
+               Fountain.Audit._unsafe_list_recent_admin(10),
+               &(&1.event_type == "admin.trial.extended" and &1.target_user_id == user.id)
+             )
+    end
+
+    test "rejects a non-numeric day count", %{conn: conn} do
+      admin = insert_admin()
+      # Trialing on purpose: the extend-trial form is only rendered for an
+      # account that has a trial, and `extend_trial/2` refuses an active one.
+      user = insert_verified_user()
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      html =
+        lv
+        |> form("#extend-trial-#{user.id}", %{"days" => "nope"})
+        |> render_submit()
+
+      assert html =~ "whole number"
+    end
+  end
+
+  describe "AdminLive.Users — toggle_comp" do
+    test "comps and un-comps an account, with audit events", %{conn: conn} do
+      admin = insert_admin()
+      user = insert_active_user()
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv
+      |> element("button[phx-value-id='#{user.id}'][phx-click='toggle_comp']")
+      |> render_click()
+
+      assert Fountain.Repo.reload!(user).subscription_status == "comped"
+
+      lv
+      |> element("button[phx-value-id='#{user.id}'][phx-click='toggle_comp']")
+      |> render_click()
+
+      assert Fountain.Repo.reload!(user).subscription_status == "canceled"
+
+      events = Fountain.Audit._unsafe_list_recent_admin(10)
+      assert Enum.any?(events, &(&1.event_type == "admin.comp.granted"))
+      assert Enum.any?(events, &(&1.event_type == "admin.comp.revoked"))
+    end
+  end
+
+  describe "AdminLive.Users — suspend (#287)" do
+    test "suspends and unsuspends with audit events and a badge", %{conn: conn} do
+      admin = insert_admin()
+      target = insert_active_user()
+      insert_sandbox(user_id: target.id, status: "ready")
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      html =
+        lv
+        |> element("button[phx-click='toggle_suspend'][phx-value-id='#{target.id}']")
+        |> render_click()
+
+      assert Fountain.Accounts.suspended?(Fountain.Repo.reload!(target))
+      assert html =~ "suspended"
+      assert html =~ "Unsuspend"
+      assert html =~ "1 sandbox(es) reaped"
+
+      html =
+        lv
+        |> element("button[phx-click='toggle_suspend'][phx-value-id='#{target.id}']")
+        |> render_click()
+
+      refute Fountain.Accounts.suspended?(Fountain.Repo.reload!(target))
+      assert html =~ "Suspension lifted"
+
+      events = Fountain.Audit._unsafe_list_recent_admin(10)
+      assert Enum.any?(events, &(&1.event_type == "admin.account.suspended"))
+      assert Enum.any?(events, &(&1.event_type == "admin.account.unsuspended"))
+    end
+
+    test "an admin cannot suspend themselves", %{conn: conn} do
+      admin = insert_admin()
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      render_click(lv, "toggle_suspend", %{"id" => admin.id})
+
+      refute Fountain.Accounts.suspended?(Fountain.Repo.reload!(admin))
+      assert render(lv) =~ "cannot suspend your own account"
+    end
+  end
+
+  describe "AdminLive.Users — search, filter, sort" do
+    # The layout renders the logged-in admin's email in the nav, so negative
+    # assertions must target a third user, never the admin.
+    test "?q= narrows the table to matching emails", %{conn: conn} do
+      admin = insert_admin()
+      needle = insert_active_user(%{email: "needle@example.com"})
+      straw = insert_active_user(%{email: "straw@example.com"})
+      conn = login_user(conn, admin)
+
+      {:ok, _lv, html} = live(conn, ~p"/admin/users?q=needle")
+
+      assert html =~ needle.email
+      refute html =~ straw.email
+    end
+
+    test "typing in the search box patches the URL and filters", %{conn: conn} do
+      admin = insert_admin()
+      needle = insert_active_user(%{email: "needle@example.com"})
+      straw = insert_active_user(%{email: "straw@example.com"})
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      html =
+        lv
+        |> form("#user-filters")
+        |> render_change(%{"q" => "needle", "status" => "", "role" => "", "verified" => ""})
+
+      assert_patch(lv, ~p"/admin/users?q=needle")
+      assert html =~ needle.email
+      refute html =~ straw.email
+    end
+
+    test "filters by subscription status", %{conn: conn} do
+      admin = insert_admin()
+      trialing = insert_active_user(%{email: "still-trialing@example.com"})
+
+      canceled =
+        Fountain.Repo.update!(
+          Ecto.Changeset.change(insert_active_user(), subscription_status: "canceled")
+        )
+
+      conn = login_user(conn, admin)
+      {:ok, _lv, html} = live(conn, ~p"/admin/users?status=canceled")
+
+      assert html =~ canceled.email
+      refute html =~ trialing.email
+    end
+
+    test "filters by verification state and badges unverified users", %{conn: conn} do
+      admin = insert_admin()
+      verified = insert_active_user(%{email: "is-verified@example.com"})
+      unverified = insert_user()
+      conn = login_user(conn, admin)
+
+      {:ok, _lv, html} = live(conn, ~p"/admin/users?verified=no")
+      assert html =~ unverified.email
+      assert html =~ "unverified"
+      refute html =~ verified.email
+    end
+
+    test "sorts by email via the header link", %{conn: conn} do
+      admin = insert_admin()
+      insert_active_user(%{email: "zzz-sort@example.com"})
+      insert_active_user(%{email: "aaa-sort@example.com"})
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      html = lv |> element("thead a", "Email") |> render_click()
+      assert_patch(lv, ~p"/admin/users?dir=asc&sort=email")
+
+      # Each email appears several times per row (confirm dialogs etc.);
+      # first-appearance order is what reflects row order.
+      assert Regex.scan(~r/(?:aaa|zzz)-sort@example\.com/, html)
+             |> List.flatten()
+             |> Enum.uniq() ==
+               ["aaa-sort@example.com", "zzz-sort@example.com"]
+    end
+
+    test "shows an empty state when nothing matches", %{conn: conn} do
+      admin = insert_admin()
+      conn = login_user(conn, admin)
+
+      {:ok, _lv, html} = live(conn, ~p"/admin/users?q=no-such-user")
+      assert html =~ "No users match."
+    end
+  end
+
+  describe "AdminLive.Users — pagination" do
+    # Bypasses register_user: 26 bcrypt hashes in one test is pure drag, and
+    # this test needs rows, not the registration pipeline.
+    defp insert_bare_users(count) do
+      for i <- 1..count do
+        Fountain.Repo.insert!(%Fountain.Accounts.User{
+          email: "bulk-#{i}-#{System.unique_integer([:positive])}@example.com",
+          password_hash: "not-a-real-hash"
+        })
+      end
+    end
+
+    test "shows 25 rows per page and preserves the page across refresh", %{conn: conn} do
+      admin = insert_admin()
+
+      # Back-date the admin so ascending order puts them firmly on page 1 —
+      # same-second inserts would otherwise tie-break on random uuid.
+      earlier = DateTime.add(DateTime.utc_now(), -60, :second) |> DateTime.truncate(:second)
+      Fountain.Repo.update!(Ecto.Changeset.change(admin, inserted_at: earlier))
+
+      insert_bare_users(26)
+      conn = login_user(conn, admin)
+
+      {:ok, lv, html} = live(conn, ~p"/admin/users?page=2&dir=asc")
+
+      # 27 users total (admin + 26): page 2 holds exactly the last two
+      assert html =~ "Page 2 of 2"
+      assert html =~ "27 accounts"
+
+      page2_emails =
+        Regex.scan(~r/bulk-\d+-\d+@example\.com/, html) |> List.flatten() |> Enum.uniq()
+
+      assert length(page2_emails) == 2
+
+      send(lv.pid, :refresh)
+      assert render(lv) =~ "Page 2 of 2"
+    end
+
+    test "prev/next links walk the pages", %{conn: conn} do
+      admin = insert_admin()
+      insert_bare_users(26)
+      conn = login_user(conn, admin)
+
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      lv |> element("a", "next →") |> render_click()
+      assert_patch(lv, ~p"/admin/users?page=2")
+
+      lv |> element("a", "← prev") |> render_click()
+      assert_patch(lv, ~p"/admin/users")
+    end
+
+    test "no pagination controls when everything fits on one page", %{conn: conn} do
+      admin = insert_admin()
+      conn = login_user(conn, admin)
+
+      {:ok, _lv, html} = live(conn, ~p"/admin/users")
+      refute html =~ "Page 1 of"
+    end
+  end
+
+  describe "AdminLive.Users — usage column" do
+    test "shows 30-day usage per user", %{conn: conn} do
+      admin = insert_admin()
+      user = insert_active_user()
+      {:ok, _} = Fountain.Billing.record_usage(user.id, "turn_started", nil, nil)
+      {:ok, _} = Fountain.Billing.record_usage(user.id, "sandbox_provisioned", nil, nil)
+
+      conn = login_user(conn, admin)
+      {:ok, _lv, html} = live(conn, ~p"/admin/users")
+
+      assert html =~ "1c · 1t · 0m"
+    end
+  end
+end
+
+defmodule FountainWeb.AdminUsersLiveErrorTest do
+  use FountainWeb.ConnCase, async: false
+  use Mimic
+
+  import Phoenix.LiveViewTest
+
+  alias Fountain.Accounts
+
+  defp insert_admin(overrides \\ %{}) do
+    user = insert_active_user(overrides)
+    {:ok, admin} = Accounts.update_user_role(user, "admin")
+    admin
+  end
+
+  describe "AdminLive.Users — toggle_admin error path" do
+    test "shows error flash when update_user_role fails", %{conn: conn} do
+      admin = insert_admin()
+      target = insert_active_user()
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      # Arity 3: the admin surface passes actor: "admin" through so the
+      # context can attribute the event it now records itself (#552).
+      stub(Accounts, :update_user_role, fn _user, _role, _opts ->
+        {:error, %Ecto.Changeset{}}
+      end)
+
+      lv |> element("button[phx-value-id='#{target.id}']", "Make admin") |> render_click()
+
+      html = render(lv)
+      assert html =~ "Failed to update role"
+    end
+  end
+
+  describe "AdminLive.Users — resync_stripe (#502)" do
+    test "adopts Stripe's state and records an audit event", %{conn: conn} do
+      admin = insert_admin()
+
+      user =
+        Fountain.Repo.update!(
+          Ecto.Changeset.change(insert_active_user(),
+            subscription_status: "past_due",
+            stripe_customer_id: "cus_lv_resync",
+            stripe_subscription_id: "sub_lv_resync"
+          )
+        )
+
+      result =
+        {:ok, %Stripe.Subscription{id: "sub_lv_resync", status: "active", trial_end: nil}}
+
+      stub(Stripe.Subscription, :retrieve, fn "sub_lv_resync" -> result end)
+      stub(Stripe.Subscription, :retrieve, fn "sub_lv_resync", _opts -> result end)
+
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      html =
+        lv
+        |> element("button[phx-value-id='#{user.id}'][phx-click='resync_stripe']")
+        |> render_click()
+
+      assert html =~ "Resynced from Stripe"
+      assert Fountain.Repo.reload!(user).subscription_status == "active"
+
+      assert Enum.any?(
+               Fountain.Audit._unsafe_list_recent_admin(10),
+               &(&1.event_type == "admin.stripe.resynced" and &1.target_user_id == user.id)
+             )
+    end
+
+    test "a user with no subscription of record gets a customer sync enqueued", %{conn: conn} do
+      admin = insert_admin()
+      user = insert_active_user()
+
+      conn = login_user(conn, admin)
+      {:ok, lv, _html} = live(conn, ~p"/admin/users")
+
+      html =
+        lv
+        |> element("button[phx-value-id='#{user.id}'][phx-click='resync_stripe']")
+        |> render_click()
+
+      assert html =~ "customer sync enqueued"
+    end
+  end
+end
